♻️Refactor: wallet 구조 변경
