+++ src/main/java/bumblebee/xchangepass/domain/user/entity/User.java
@@ -2,7 +2,7 @@ package bumblebee.xchangepass.domain.user.entity;
 
 import bumblebee.xchangepass.domain.ExchangeTransaction.entitiy.ExchangeTransaction;
 import bumblebee.xchangepass.domain.user.entity.value.*;
-import bumblebee.xchangepass.domain.wallet.entity.Wallet;
+import bumblebee.xchangepass.domain.wallet.wallet.entity.Wallet;
 import jakarta.persistence.*;
 import lombok.Builder;
 import lombok.Getter;
