+++ src/main/java/bumblebee/xchangepass/domain/user/service/UserRegisterService.java
@@ -3,8 +3,7 @@ package bumblebee.xchangepass.domain.user.service;
 import bumblebee.xchangepass.domain.user.dto.request.UserRegisterRequest;
 import bumblebee.xchangepass.domain.user.entity.User;
 import bumblebee.xchangepass.domain.user.repository.UserRepository;
-import bumblebee.xchangepass.domain.user.service.NicknameGenerator;
-import bumblebee.xchangepass.domain.wallet.service.WalletService;
+import bumblebee.xchangepass.domain.wallet.wallet.service.WalletService;
 import bumblebee.xchangepass.global.error.ErrorCode;
 import bumblebee.xchangepass.global.exception.CommonException;
 import bumblebee.xchangepass.global.util.DuplicateKeyExceptionHandler;
