+++ src/main/java/bumblebee/xchangepass/domain/wallet/balance/repository/WalletBalanceRepository.java
@@ -1,15 +1,13 @@
-package bumblebee.xchangepass.domain.walletBalance.repository;
+package bumblebee.xchangepass.domain.wallet.balance.repository;
 
-import bumblebee.xchangepass.domain.walletBalance.entity.WalletBalance;
+import bumblebee.xchangepass.domain.wallet.balance.entity.WalletBalance;
 import jakarta.persistence.LockModeType;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Lock;
-import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
-import java.math.BigDecimal;
 import java.util.Currency;
 import java.util.List;
 import java.util.Optional;
