--- src/main/java/bumblebee/xchangepass/domain/wallet/dto/request/WalletCreateRequest.java
@@ -1,6 +1,0 @@
-package bumblebee.xchangepass.domain.wallet.dto.request;
-
-public record WalletCreateRequest(
-        Long userId
-) {
-}
