+++ src/main/java/bumblebee/xchangepass/domain/wallet/wallet/dto/request/WalletCreateRequest.java
@@ -1,0 +1,6 @@
+package bumblebee.xchangepass.domain.wallet.wallet.dto.request;
+
+public record WalletCreateRequest(
+        Long userId
+) {
+}
