+++ src/main/java/bumblebee/xchangepass/domain/wallet/wallet/entity/Wallet.java
@@ -1,19 +1,16 @@
-package bumblebee.xchangepass.domain.wallet.entity;
+package bumblebee.xchangepass.domain.wallet.wallet.entity;
 
 import bumblebee.xchangepass.domain.card.entity.Card;
 import bumblebee.xchangepass.domain.user.entity.User;
-import bumblebee.xchangepass.domain.walletBalance.entity.WalletBalance;
+import bumblebee.xchangepass.domain.wallet.balance.entity.WalletBalance;
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
 import java.time.LocalDateTime;
 import java.util.List;
-
-import static lombok.AccessLevel.PROTECTED;
 
 @Entity
 @Table(name = "wallet")
@@ -23,23 +20,23 @@ import static lombok.AccessLevel.PROTECTED;
 public class Wallet {
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "wallet_id", nullable = false)
-    public Long walletId;
+    private Long walletId;
 
     @Column(name = "wallet_password", nullable = false)
-    public String walletPassword;
+    private String walletPassword;
 
     @OneToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "user_id")
-    public User user;
+    private User user;
 
     @OneToMany(mappedBy = "wallet", fetch = FetchType.LAZY, orphanRemoval = true, cascade = CascadeType.ALL)
-    List<Card> cards;
+    private List<Card> cards;
 
     @OneToMany(mappedBy = "wallet", fetch = FetchType.LAZY, orphanRemoval = true, cascade = CascadeType.ALL)
-    List<WalletBalance> walletBalances;
+    private List<WalletBalance> walletBalances;
 
     @CreatedDate
-    public LocalDateTime walletCreatedAt;
+    private LocalDateTime walletCreatedAt;
 
     public Wallet(User user, String walletPassword) {
         this.user = user;
