+++ src/main/java/bumblebee/xchangepass/domain/wallet/wallet/repository/LockRepository.java
@@ -1,7 +1,6 @@
-package bumblebee.xchangepass.domain.wallet.repository;
+package bumblebee.xchangepass.domain.wallet.wallet.repository;
 
-import bumblebee.xchangepass.domain.wallet.entity.Wallet;
-import org.hibernate.cache.spi.support.AbstractReadWriteAccess;
+import bumblebee.xchangepass.domain.wallet.wallet.entity.Wallet;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
